fix(email): address review findings on unsubscribe
- remove defensive .catch() in handleComplaint(), let errors propagate
- increase unsubscribe HMAC token from 64-bit to 128-bit
- fix unsafe `as string[]` cast with proper type guard
- extract shared UnsubscribeFooter component
