+++ turbo/apps/web/src/lib/email/templates/inbound-error.tsx
@@ -1,12 +1,5 @@
-import {
-  Html,
-  Head,
-  Body,
-  Container,
-  Text,
-  Link,
-  Hr,
-} from "@react-email/components";
+import { Html, Head, Body, Container, Text, Hr } from "@react-email/components";
+import { UnsubscribeFooter } from "./unsubscribe-footer";
 
 interface InboundErrorEmailProps {
   errorMessage: string;
@@ -25,13 +18,7 @@ export function InboundErrorEmail({
           <Text style={errorStyle}>{errorMessage}</Text>
           <Hr style={hrStyle} />
           <Text style={signatureStyle}>VM0</Text>
-          {unsubscribeUrl && (
-            <Text style={unsubscribeFooterStyle}>
-              <Link href={unsubscribeUrl} style={linkStyle}>
-                Unsubscribe
-              </Link>
-            </Text>
-          )}
+          <UnsubscribeFooter unsubscribeUrl={unsubscribeUrl} />
         </Container>
       </Body>
     </Html>
@@ -70,15 +57,4 @@ const signatureStyle = {
   fontWeight: "600" as const,
   color: "#374151",
   margin: "0",
-};
-
-const unsubscribeFooterStyle = {
-  fontSize: "12px",
-  color: "#9ca3af",
-  margin: "16px 0 0",
-};
-
-const linkStyle = {
-  color: "#2563eb",
-  textDecoration: "underline",
 };
