+++ turbo/apps/web/src/lib/email/templates/schedule-failed.tsx
@@ -7,6 +7,7 @@ import {
   Link,
   Hr,
 } from "@react-email/components";
+import { UnsubscribeFooter } from "./unsubscribe-footer";
 
 interface ScheduleFailedEmailProps {
   agentName: string;
@@ -34,13 +35,7 @@ export function ScheduleFailedEmail({
               Audit
             </Link>
           </Text>
-          {unsubscribeUrl && (
-            <Text style={unsubscribeFooterStyle}>
-              <Link href={unsubscribeUrl} style={linkStyle}>
-                Unsubscribe
-              </Link>
-            </Text>
-          )}
+          <UnsubscribeFooter unsubscribeUrl={unsubscribeUrl} />
         </Container>
       </Body>
     </Html>
@@ -85,12 +80,6 @@ const footerStyle = {
   fontSize: "13px",
   color: "#6b7280",
   margin: "0",
-};
-
-const unsubscribeFooterStyle = {
-  fontSize: "12px",
-  color: "#9ca3af",
-  margin: "16px 0 0",
 };
 
 const linkStyle = {
